+ Power Sub-Station now uses 1% of average eu/t. $ Possible fix to dynamo hatches not filling properly. + Added a much better tooltip to the Power Sub-Station.

+++ src/Java/gtPlusPlus/core/util/math/MathUtils.java
@@ -144,7 +144,7 @@ public class MathUtils {
 		return roundOff;
 	}
 	
-	public static int findPercentageOfInt(int input, int percentage){
+	public static int findPercentageOfInt(long input, float percentage){
 		return (int)(input*(percentage/100.0f));
 	}
